Ver 2.2
Added "Worst Case" checkbox. Dealt with what happens when some of the
dropdowns are empty, so no errors are thrown when they are updated in
code one at a time and the interrupt tries to retrieve data with half of
them blank.  Made "Clear All" function clear dropboxes and checkbox.